fix: rename basic example back to original run-in-child-context name
- Rename basic-run-in-child-context.ts back to run-in-child-context.ts
- Keep virtual-run-in-child-context.ts with new name
- Update CloudFormation template with correct handler names
- Fix integration test module resolution issues
